creates variable_gen.rb
this necessary after TSE feedbback, classes should be in different files
